requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.30.0
langchain>=0.0.292
langchain-openai>=0.0.2
openai>=1.3.0
sqlite3-api>=0.1.0
bcrypt>=4.1.0
pyjwt>=2.8.0
cryptography>=41.0.0
pydantic>=2.5.0
bleach>=6.1.0

